# api/routers/alerts.py
"""
告警管理路由

作者: Backend Team
职责: 告警查询、确认、统计
"""

from typing import List, Optional
from datetime import datetime
from fastapi import APIRouter, Query
from pydantic import BaseModel

router = APIRouter()


class Alert(BaseModel):
    """告警模型"""
    alert_id: str
    rule_id: str
    rule_name: str
    severity: str  # critical, warning, info
    message: str
    timestamp: datetime
    acknowledged: bool
    acknowledged_by: Optional[str] = None
    acknowledged_at: Optional[datetime] = None


class AcknowledgeRequest(BaseModel):
    """确认告警请求"""
    operator: str
    comment: Optional[str] = None


@router.get("/", response_model=List[Alert])
async def get_alerts(
    severity: Optional[str] = Query(None, enum=["critical", "warning", "info"]),
    acknowledged: Optional[bool] = None,
    start_time: Optional[datetime] = None,
    end_time: Optional[datetime] = None,
    limit: int = Query(50, ge=1, le=200)
):
    """
    获取告警列表
    
    支持筛选条件:
    - severity: 严重级别
    - acknowledged: 是否已确认
    - start_time/end_time: 时间范围
    """
    # TODO: 从数据库查询
    return []


@router.get("/active")
async def get_active_alerts():
    """获取当前活动告警"""
    # TODO: 从 RuleEngine 获取
    return {
        "alerts": [],
        "count": 0,
        "critical_count": 0,
        "warning_count": 0
    }


@router.post("/{alert_id}/acknowledge")
async def acknowledge_alert(alert_id: str, request: AcknowledgeRequest):
    """确认告警"""
    return {
        "alert_id": alert_id,
        "acknowledged": True,
        "acknowledged_by": request.operator,
        "acknowledged_at": datetime.now().isoformat()
    }


@router.get("/statistics")
async def get_alert_statistics(
    days: int = Query(7, ge=1, le=90)
):
    """
    获取告警统计
    
    - **days**: 统计天数
    """
    return {
        "total": 45,
        "critical": 5,
        "warning": 25,
        "info": 15,
        "acknowledged": 40,
        "unacknowledged": 5,
        "by_day": []
    }