# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Miaota Industrial Agent API
  description: |
    工业智能监控与诊断系统API
    
    ## 功能模块
    - **数据采集**: PLC数据采集与管理
    - **设备管理**: 设备状态监控与配置
    - **告警管理**: 告警规则与通知
    - **数据分析**: 异常检测与趋势分析
    - **知识库**: 故障诊断与知识检索
    - **报表**: 数据导出与可视化
    
  version: v1.0.0-beta1
  contact:
    name: Miaota Team
    email: team@miaota.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发环境
  - url: https://api.miaota.local/api/v1
    description: 生产环境
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /devices:
    get:
      tags: [设备管理]
      summary: 获取设备列表
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [s7, modbus]
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, error]
      responses:
        '200':
          description: 设备列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'

    post:
      tags: [设备管理]
      summary: 创建设备
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
      responses:
        '201':
          description: 创建成功
...
  /alerts:
    get:
      tags: [告警]
      summary: 获取告警列表
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: 告警列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
  /alerts/{alert_id}/acknowledge:
    post:
      tags: [告警]
      summary: 确认告警
      parameters:
        - name: alert_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        '200':
          description: 确认成功
...
  /rules:
    get:
      tags: [告警]
      summary: 获取告警规则
      responses:
        '200':
          description: 规则列表

    post:
      tags: [告警]
      summary: 创建告警规则
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRule'
      responses:
        '201':
          description: 创建成功

  # ============ 数据分析 ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API密钥认证
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [s7, modbus]
        host:
          type: string
        port:
          type: integer
        status:
          type: string
          enum: [online, offline, error]
        last_seen:
          type: string
          format: date-time
        tag_count:
          type: integer
    DeviceTag:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        data_type:
          type: string
        unit:
          type: string
        description:
          type: string
    DeviceCreateRequest:
      type: object
      required: [name, type, host, port]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [s7, modbus]
        host:
          type: string
        port:
          type: integer
        rack:
          type: integer
        slot:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/DeviceTag'
...
    Alert:
      type: object
      properties:
        id:
          type: string
        rule_id:
          type: string
        rule_name:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        message:
          type: string
        device_id:
          type: string
        tag:
          type: string
        value:
          type: number
        threshold:
          type: number
        status:
          type: string
          enum: [active, acknowledged, resolved]
        created_at:
          type: string
          format: date-time
        acknowledged_by:
          type: string
        acknowledged_at:
          type: string
          format: date-time
    AlertRule:
      type: object
      properties:
        rule_id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        condition:
          type: object
          properties:
            type:
              type: string
            tag:
              type: string
            operator:
              type: string
            value:
              type: number
        severity:
          type: string
        message:
          type: string
        suppression_window_minutes:
          type: integer

    # 数据相关
...